Module my_app::game

Enums

Direction 🔒

Used to get directions, and nothing else !

Constants

BOOL_OPTIONS_NUMBER

The number of Boolean Options used. Public so menu() can use it.

CASE_SIZE 🔒

The size of a grid case. Everything is linked to this value.

DARK_GREEN 🔒

GRAY 🔒

LIGHT_GRAY 🔒

LIGHT_GREEN 🔒

MAX_ARRAY_SIZE 🔒

No array in this program can exceed this size ; it can be used to set the size.

MAX_HEIGHT 🔒

The max y value on the grid

MAX_WIDTH 🔒

THe max x value on the grid

Functions

draw_box 🔒

Draws a box (case) of the grid

draw_snake 🔒

Draws the snake !

draw_terrain 🔒

Draws the entire terrain

draw_terrain_box 🔒

Draws a terrain box (a box with the right color)

game

The entire game is here.

get_point_from_dir 🔒

Return a Point from a Point and a Direction, None if the Point is non-valid

get_random_point 🔒

Give a random Point, corresponding to a position on the game grid

is_in_snake 🔒

Check if a given Point is in a Deque (the snake)

is_in_walls 🔒

Check if a given Point is in a Vec (the walls)

start

Menu, Options and Game start