Module my_app::menu

Structs

MyOption

An option

Enums

CursorPos 🔒

Used to symbolise the cursor position

Constants

EXIT_POS 🔒

The position of EXIT_TXT

EXIT_TXT 🔒

The Exit option text

FADING_TIME 🔒

The duration of every fadings in the menus

OPTIONS_POS 🔒

The position of OPTIONS_TXT

OPTIONS_TXT 🔒

The Options option text

PAUSE_RECT_SIZE 🔒

How big is the rectangle on the top & bottom, *2 for left and right.

RESUME_TXT 🔒

The text of the Start/Resume option

SPACE_BETWEEN_LINES 🔒

For the options menu, the space between each line.

START_POS 🔒

The position of START_TXT

START_TXT 🔒

The Start option text

XPOS_NAMES 🔒

Where in the x coordinate will the names of the option be placed

XPOS_VALUES 🔒

Where in the x coordinate will the values (option.value.1) be placed

Functions

draw_options_selection 🔒

Draws the line corresponding to the given option value

draw_pause_selection_string 🔒

Draws the line corresponding to the given CursorPos (for pause menu)

draw_selection_string 🔒

Draws the line corresponding to the CursorPos

menu

Creates a fully fonctional start menu, with Options !

options 🔒

Create a fully fonctional option menu, which changes directly the options values. (no option return)

pause_menu

Creates a fully foncional pause menu