Module my_app::utils

Constants

LARGE_CHAR_HEIGHT

LARGE_CHAR_WIDTH

SMALL_CHAR_HEIGHT

SMALL_CHAR_WIDTH

Functions

draw_centered_string

Draws the given string centered in the x coordinate

fading

Make the screen fade to 0 brightness, then back to original brightness

fill_screen

Fills the screen with the given Color

get_centered_text_left_coordo

Returns the left x coordinate that centers the given string

randint

Gives a random integer between a and b (included)